--- redisai/_util.py
@@ -1,10 +1,0 @@
-import six
-
-
-def to_string(s):
-    if isinstance(s, six.string_types):
-        return s
-    elif isinstance(s, bytes):
-        return s.decode('utf-8')
-    else:
-        return s  # Not a string we care about
